Add automatic and dynamic completion for aliases
Adds a compalias() shell function that creates an alias and then
applies the completion properties of the true command onto the
alias. This includes both copying over any existing completion
that exists when the shell is loaded, and also creating a
completion file in ~/.local/share that will dynamically load
completions just like the true command.

Both elements are required in situations like fzf, which generate
completion commands on startup but also use dynamic completion.
aliases.sh is now sourced after fzf.sh in order to catch this.

kubectl completion was also moved to a separate file that can will
be sourced dynamically.