Add new setup script
Beginning a new setup script to manage dotfiles installation. The new
script will work in stages, first calculating the desired state, then
checking everything to determine what needs to be done, and then running
jobs. This is instead of doing checks at the same time as running jobs.
This first draft includes displays for desired state and incorrect
state, but is not yet creating jobs.