Move python startup code all into one file
I don't need two different files for python startup. Instead I can
just use one file that handles the cases of both being in or not in
a virtual environment. One file makes it easier to manage, as before
this the venv startup wasn't using the custom history file location.

There are some small improvements to `venv_management.py` in this
commit.