Use lazy.nvim for package management
I'm a little hesitant to make this change, due to lazy.nvim not being
compatiable with vim's builtin package management system, so it's all
or nothing. But ultimately I am interested in using the lazy loading,
and it makes sense to do this now before moving forward with any more
package updating, so I can just update and transition to lazy load at
the same time.

The packadds in vimrc are now only useful for vanilla vim, and init.lua
has to specify all the package in the vimrc on its own as well.