Update and lazy load nvim-treesitter
The plugin is being lazy loaded off the 'VeryLazy' autocmd, which
causes the highlighting to lag slightly on load. I will reevaluate this
once my startuptime is reduced further to see if it is still an issue.

Removing the playground plugin because it has been deprecated and the
functionality is in neovim core now.