Update nvim-tree.lua
The plugin has changed its api a bit over this update. Keymaps
are now declared in an `on_attach` function, and I needed to switch
to `tree.open()` for the open in place part of `vinegar_open`. On
the plus side, the update now handles alternate files properly when
using netrw style navigation.