Insert tab when no completions available
For a while now, I have been frustrated with not being able to tab
at the end of a word. I think the option to open the completion menu
is still valid, for instanec right when entering insert mode, but
repeatedly pressing tab with nothing happening doesn't feel right.

With this change, tab will still attempt to open the completion menu,
but if no suggestions pop up, an actual tab will be inserted. This
feels like a good middle ground to keep the old benefits and fix my
irks.