Refactor statusline to properly get values
My initial implementation of the lua statusline in the previous commit
had an issue where the `vim.b` values were all being evaluated in the
buffer the cursor was currently in. This meant that the path,
diagnostics and cursor position pieces were all changing based on the
focused buffer.

To fix this, I've refactored the statusline to use `%{}` evaluations
for each component, rather than just being a single eval of a lua
function.