Add --first-parent flag to git aliases
I've been frustrated by this issue for a long time where my `git last`
and `git showlast` aliases don't properly show diffs/stats on merge
commits. The missing piece was thie `--first-parent` flag, which tells
git to compute the diff using the first parent of the merge, which is
what I probably want in 99% of situations.