Create vim utilities for resolving merge conflicts
Leveraging fugitive, I've created some keymaps and commands for
assisting in resolving merge conflicts in vim:
* A `:Resolve` command, which calls `:Gvdiffsplit` to set up a
three-way diff, determines whether a merge or rebase is active, and
sets nomodifiable and readonly flags on the buffers
* Keymaps to easily select a version of the conflict from the left or
right side
* Keymaps to navigate between merge conflict markers - adapted from
vim-unimpaired
* Custom statusline messages that better indicate what each side of
a merge conflict represent
* `:Finish` command that closes the fugitive buffers and turns off
readonly mode.

`fugitive` was added as a difftool in my gitconfig to automatically
open vim with `:Resolve` when resolving conflicts.

Additionally;
* Add the option diffopts+=linematch:60, which changes the diff
algorithm used by vim to create more compact diffs
* Drop the `VeryLazy` event from nvim-treesitter's loading condition.
`VeryLazy` runs after commands passed in the with the -c flag. So when
opening vim with git mergetool, treesitter loads after `:Resolve`,
which was messing with folding.