perf(submodules) only fetch partial repo history
Passing the `--depth=10` flag to git submodule update so that it only
grabs the latest 10 commits instead of the whole history of each
repository. This saves close to 100MB which is especially nice when
compressing dotfiles for other transfers.

Also trueing up .gitmodules to name each submodule which the current
path.