feat(setup) Add InstallerSpec class
The InstallerSpec provides a way to further customize how a package
is installed with a particular manager. The `package_name` field lets
you define a package name other than the name of the resource, and
the `post_script` field is intended to allow further custom script
steps after installation.

`post_script` is not implemented in this commit.