refactor(setup) Remove check_results variable
Removed the check_results class variable and instead add a state
field to every manager instance. This is a little easier to manage and
also makes it easy to check the status of individual resources. Also
add a set_status() method to the Manager ABC to make it easier to
check statuses of a small set of resources instead of an entire type
at once.