feat(vim, shell) Rewrite colorscheme features
This is a more-or-less complete rewrite of all of my colorscheme
management code. The biggest changes is removing base16-shell and
base16-vim in favor of using tinted-theming/schemes directly and
implementing the actual application of the schemes on my own.

For vim, the new `:Color` command acts as a replacement for
`:colorscheme` that uses the tinted-theming yaml schemes and properly
parses them and applies them. The command supports argument completion.
All highlight commands are now included directly in the dotfiles repo,
which also gives me greater control on customizing them and adding
additional highlight groups as needed. I've removed pinnacle entirely
now too.

In the shell, I've replaced `color.sh` with `term_color.py`.
`term_color.py` uses the same logic and escape sequences as the
original base16-shell files plus the custom code that was in
`color.sh`. I switched to python because it was simpler for parsing the
yaml. The script is symlinked to a `color` command in `~/.local/bin`.