fix(vim) Utilize default rust ftplugin
At some point in the past I disabled the default rust ftplugin. I
suspect I did this because it was overriding my textwidth and
indent options. But I've now noticed that it also disabled the setting
of commentstring, so commenting was working. I am reenabling the
plugin and just disabling the style overrides, hopefully that doesn't
mess up my environment.