feat(vim) Replace vimwiki with obsidian.nvim
I've been using Obsidian for my notes on mobile for a while and I
intend to continue using it. I am not interested in too many features
of either tool but I think ultimately obsidian.nvim makes more sense
moving forward as it is more active recently and neovim-specific. It
is missing some features from vimwiki that I liked but they're minor
and I can deal with that in the future.