chore: Remove vim-tmux-navigator
I kept this around solely as an alternative for tmux.nvim specifically
when using vanilla vim. I feel that this is such a minor use case now
that it doesn't warrant keeping. In the rare cases where I would need
to use vanilla vim I will just just the the builtin keybindings.