perf(setup): Fetch NPM exe version's immediately
I wanted to improve the time it takes to check that everything is up to
date. I found that the major source of time loss comes from multiple
NPM-installed language servers that don't have any `--version` command.
As a result, I fallback to looking up the version directly in the
`package.json`. The time loss comes from trying and failing all the
different variations of a version flag. To speed this up, I now skip
the version flag attempts and use the `package.json` immediately if
`version_cmd` is set to "NPM". This is a bit of a hack but I'm happy
enough with it for now.

The other noteworthy timeloss comes from pylint. `pylint --version`
works, but does not return as quickly as most commands. Yet another
reason to migrate away from pylint. I am using a pretty old version, I
will update soon and see if that helps.

With all the improvements (including removing pylint), `mysetup` runs
in about ~0.5 seconds. Before the change it regularly took over a
second.