chore: Update python development tools
In 3.0.0 pylint changed the default behavior to allow variable names
of length 1 or 2, so I removed my override in .pylintrc.

As investigated in commit 6fe0034, I looked at the performance change
of `pylint --version`. I think it is a little faster, but still not
as fast as most version commands.