fix(vim) Use builtin vim.fs.root to determine root_dir
My `get_git_root` was pretty hacky and was failing when not in a
git repo. While trying to fix it I found `vim.fs.root` which does
what I need and is much cleaner. I also added a fallback to use
the immediate parent of the current file if not in a git repo.