feat(setup): Support multiple dependencies on jobs
Using futures, jobs can now be dependent on multiple parents. All but
one parent will mark a future as done when complete. The child is then
put in a wrapper that is dependent on the last parent, and that wrapper
waits for all the futures to finish before calling the child.