refactor: Handle job ordering with futures
I rewrote `build_tree()` to manage job dependencies with only futures,
instead of making jobs children of other jobs. All jobs are now at the
"root" level and are gathered at the same time. Jobs with dependencies
have a set of futures that they await on, and these futures are set
by the parents when they complete.

The big loss here is that `print_job_tree()` can no longer easily
display a tree, since the jobs are actually constructed in a tree
structure. I did try to build a tree structure alongside the future
creation to be used only for display, but that ended up being
non-trivial. For now `print_job_tree()` just prints a count of parents
and children for every job.