refactor(setup): Bump min python version to 3.9
Python3.8 end of life has already passed. Ubuntu20.04, which uses
python3.8 as a base is also reaching end of life in a couple of months.
My main environment is now on Ubuntu22.04, so I feel it's now
reasonable to bump to 3.9.

This allowed me to clean up a bunch of types in factor of Generic
alias types, and use the new string `removeprefix()` method.