feat(vim) Add CSS LSP and linter
I already have the CSS language server installed as part of
`vscode-langservers-extracted`, but now I track it separately and have
it set up in vim as `cssls`. I also added `stylelint` as a
linter/formatter for CSS. It's a little slow, but oh well. I may need to
tune the rules as I use it more.

Ticket#: DOT-12