feat(setup) Add `only_on` and `exit_code` to specs
`only_on` can apply to any spec and indicates that a spec can only be
managed on certains systems, i.e. "Linux" or "Windows". When on a
non-compliant system, generating the resource will fail and it will not
be checked.

`exit_code` is for command resources only. It specifies what the
expected exit code for the `check_script` is. I am using this instead of
using `!` to check for a nonzero exit_code. This way should be more
adaptable.