refactor: Rewrite clipin and clipout in python
Rewriting the clipboard commands in python has a few benefits. First,
they are more portable and now easily work on both Windows and Linux.
Second, python functions and string manipulation make it easier to
handle strange file encodings and deal with the extra newline that
`powershell` is inserting in the paste.

To work on Windows, both the symlink and the original file need proper
file extensions. This allows me to take advantage of Window's file
associations to call the commands like an executable.