feat(setup) Platform-specific installs for archives
This commit modifies the Tar and Zip installers to support installs only
on specific systems. Systems are keyed with the same names used by
build-hole. The installers can now include a `url` map that will pull
specific files based on the system, or fail if the target system is
not supported.