feat(nushell) Improve right-prompt format
This shortens the timestamp in nushell's right prompt by removing the
leading digits of the year and removing the AM/PM marker. I did this to
make it more concise and consistent with how I use times elsewhere.

I've also changed the table style to "none", which makes outputs much
more concise.