feat(setup) Manage tree-sitter CLI
With the rewrite of nvim-treesitter, I now need the tree-sitter CLI in
order to install new parsers.

Another effect of the rewrite is that nvim-treesitter now supports
multiple language installs in a single command, so I can refactor the
parse install system to run as a single job and get rid of the dummy
`nvim_init` job.