feat(vim) Replace command-t with snacks.nvim
It's a sad day, but I've decided to remove command-t in favor of
snacks.nvim. I did some basic testing and comparisons, and snacks seems
to match command-t for filtering as far as I need it. Moving everything
to a single picker engine will make further improvements easier to
implement.