fix(vim) Change keymaps for window resizing
My old horizontal resize cleanup key ,`<leader>h=`, conflicted
with help search (`<leader>h`). This was causing the help search to take
a while to startup becaused it was waiting for the keymap timeout.

I've updated the map to now be `<leader>=h` and updated the other
similar maps to follow the same format.